requirements.txt
# Quot PSE — Python runtime dependencies
#
# Pinned to the versions that Sprint 1-4 work has been validated against.
# Update via pip-compile / manual pinning when bumping; do not use
# unpinned ranges in production.

# ── Core framework ─────────────────────────────────────────────────────
Django==4.2.28
django-tenants==3.7.0
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.6.0
django-simple-history==3.7.0

# ── Database ───────────────────────────────────────────────────────────
psycopg2-binary==2.9.9

# ── Auth / security ────────────────────────────────────────────────────
# (add django-otp + qrcode when MFA ticket lands)

# ── Background / scheduling (optional) ─────────────────────────────────
# django-celery-beat is imported behind a try/except in settings.py;
# include it when the tenant needs scheduled jobs.
# django-celery-beat==2.7.0
# celery==5.4.0
# redis==5.2.0

# ── File / report rendering ────────────────────────────────────────────
openpyxl==3.1.5
weasyprint==63.1    # for PDF export of IPSAS statements (when the
                    # export ticket lands; harmless to preinstall)

# ── Development + test ─────────────────────────────────────────────────
# These are harmless in production images but mostly used by CI. Keep
# them pinned so pytest-django and the unit tests produce the same
# results everywhere.
pytest==7.4.4
pytest-django==4.9.0
pytest-cov==5.0.0
ruff==0.7.4
bandit==1.7.10
